pyproject.toml
[project]
name = "bels-shop"
version = "0.1.0"
description = ""
authors = [
    {name = "KirillBels",email = "k.belskiy@icloud.com"}
]
readme = "README.md"
requires-python = "3.11.9"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.1,<4.0.0)",
    "pydantic[email] (>=2.11.3,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "argon2-cffi (>=23.1.0,<24.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "ruff (>=0.11.7,<0.12.0)",
    "celery (>=5.5.2,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "fastapi-cache2 (==0.2.2)",
    "requests (>=2.32.3,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pytest (>=8.3.5,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
uvicorn = {extras = ["standard"], version = "^0.34.2"}
pytest-dotenv = "^0.5.2"
pytest-asyncio = "^0.26.0"


[tool.ruff]
line-length = 80

